refactor(pkg/swagger): Loader interface
